Implemented timecard and wage system (#24)
* First version of timecard and wage system

* Changed money system to use subunits instead of floats

* Formatted

* Fixed more problems

* Fixed additional issues

* Formatted

* Timezone fixes

+++ internal/domain/billing.go
@@ -32,9 +32,9 @@ type ClaimLineItem struct {
 	Surfaces         []ToothSurface `json:"surfaces,omitempty"`
 	ADACode          string         `json:"ada_code"`
 	Description      string         `json:"description"`
-	Fee              float64        `json:"fee"`
-	InsuranceAllowed float64        `json:"insurance_allowed,omitempty"`
-	PatientPortion   float64        `json:"patient_portion,omitempty"`
+	Fee              int64          `json:"fee"`
+	InsuranceAllowed int64          `json:"insurance_allowed,omitempty"`
+	PatientPortion   int64          `json:"patient_portion,omitempty"`
 }
 
 // Claim represents an insurance claim for one or more procedures.
@@ -55,8 +55,8 @@ type Claim struct {
 }
 
 // TotalFee sums all line item fees on the claim.
-func (c *Claim) TotalFee() float64 {
-	var total float64
+func (c *Claim) TotalFee() int64 {
+	var total int64
 	for _, item := range c.LineItems {
 		total += item.Fee
 	}
@@ -68,7 +68,7 @@ type Payment struct {
 	ID        string        `json:"id"`
 	PatientID string        `json:"patient_id"`
 	ClaimID   string        `json:"claim_id,omitempty"` // nullable — payment may not be tied to a specific claim
-	Amount    float64       `json:"amount"`
+	Amount    int64         `json:"amount"`
 	Method    PaymentMethod `json:"method"`
 	Date      string        `json:"date"` // stored as YYYY-MM-DD
 	Notes     string        `json:"notes,omitempty"`
@@ -77,17 +77,17 @@ type Payment struct {
 
 // PatientBalance is a computed summary DTO for a patient's outstanding balance.
 type PatientBalance struct {
-	PatientID   string  `json:"patient_id"`
-	TotalBilled float64 `json:"total_billed"`
-	TotalPaid   float64 `json:"total_paid"`
-	Outstanding float64 `json:"outstanding"`
+	PatientID   string `json:"patient_id"`
+	TotalBilled int64  `json:"total_billed"`
+	TotalPaid   int64  `json:"total_paid"`
+	Outstanding int64  `json:"outstanding"`
 }
 
 // BundleItemTemplate is a CDT line item template stored inside a TreatmentBundle.
 type BundleItemTemplate struct {
-	ADACode     string  `json:"ada_code"`
-	Description string  `json:"description"`
-	DefaultFee  float64 `json:"default_fee"`
+	ADACode     string `json:"ada_code"`
+	Description string `json:"description"`
+	DefaultFee  int64  `json:"default_fee"`
 }
 
 // TreatmentBundle is a practice-wide reusable template of CDT procedures.
@@ -99,7 +99,7 @@ type TreatmentBundle struct {
 	Name        string               `json:"name"`
 	Description string               `json:"description,omitempty"`
 	Items       []BundleItemTemplate `json:"items"`
-	TotalFee    float64              `json:"total_fee"`
+	TotalFee    int64                `json:"total_fee"`
 	CreatedAt   time.Time            `json:"created_at"`
 	UpdatedAt   time.Time            `json:"updated_at"`
 }
@@ -110,8 +110,8 @@ type ProcedureCode struct {
 	Code         string      `json:"code"`
 	Category     string      `json:"category"`
 	Description  string      `json:"description"`
-	DefaultFee   float64     `json:"default_fee"`
-	EffectiveFee float64     `json:"effective_fee,omitempty"`
+	DefaultFee   int64       `json:"default_fee"`
+	EffectiveFee int64       `json:"effective_fee,omitempty"`
 	IsActive     bool        `json:"is_active"`
 }
 
@@ -121,6 +121,6 @@ type FeeSchedule struct {
 	CountryCode CountryCode `json:"country_code"`
 	Code        string      `json:"code"`
 	ProviderID  string      `json:"provider_id,omitempty"`
-	CustomFee   float64     `json:"custom_fee"`
+	CustomFee   int64       `json:"custom_fee"`
 	UpdatedAt   time.Time   `json:"updated_at"`
 }
