+++ internal/storage/sqlite/migrations/20260808000001_billing_schema.sql
@@ -26,7 +26,7 @@ CREATE TABLE IF NOT EXISTS payments (
     id TEXT PRIMARY KEY,
     patient_id TEXT NOT NULL,
     claim_id TEXT DEFAULT '',
-    amount REAL NOT NULL,
+    amount INTEGER NOT NULL,
     method TEXT NOT NULL DEFAULT 'cash',
     date TEXT NOT NULL,
     notes TEXT DEFAULT '',
@@ -44,7 +44,7 @@ CREATE TABLE IF NOT EXISTS treatment_bundles (
     name TEXT NOT NULL,
     description TEXT DEFAULT '',
     items TEXT DEFAULT '[]',
-    total_fee REAL NOT NULL DEFAULT 0.0,
+    total_fee INTEGER NOT NULL DEFAULT 0,
     created_at DATETIME NOT NULL,
     updated_at DATETIME NOT NULL
 );
